Correction Ajout nouveau forfait

+++ src/app/table-forfaits/table-forfaits.component.ts
@@ -3,7 +3,6 @@ import { MatDialog } from '@angular/material/dialog';
 import { DialogFormulaireForfaitsComponent } from '../components-utilitaires/dialog-formulaire-forfaits/dialog-formulaire-forfaits.component';
 import { ForfaitService } from '../services/forfait.service';
 import { Forfait } from '../interfaces/forfait';
-import { Hotel } from '../interfaces/hotel';
 
 @Component({
   selector: 'app-table-forfaits',
@@ -27,8 +26,10 @@ export class TableForfaitsComponent implements OnInit {
   }
 
   onDelete(forfait: Forfait): void {
-    this.forfaitService.deleteForfait(forfait._id)
-      .subscribe(result => this.forfaits = this.forfaits.filter(p => p !== forfait));
+    if(forfait._id){
+      this.forfaitService.deleteForfait(forfait._id)
+        .subscribe(result => this.forfaits = this.forfaits.filter(p => p !== forfait));
+    }
   }
 
   onSelect(forfait?: Forfait) {
@@ -41,14 +42,14 @@ export class TableForfaitsComponent implements OnInit {
         prix: 0,
         rabais: 0,
         vedette: false,
-        hotel:<Hotel>{
+        hotel: {
           nom: "",
           coordonnees: "",
           nombreEtoiles: 0,
           nombreChambres: 0,
-          caracteristiques: [""]
+          caracteristiques:[]
         }
-      }; // {} as Forfait;
+      };
     } else { 
       this.selectedForfait = forfait;
     }
